# docs/swagger.yaml
definitions:
...
  request.Login:
    properties:
      mobile:
        type: string
      password:
        type: string
    required:
    - mobile
    - password
    type: object
...
  response.Response:
    properties:
      data:
        description: 数据
      error_code:
        description: 自定义错误码
        type: integer
      message:
        description: 信息
        type: string
    type: object
info:
  contact: {}
  description: 本文档主要是提供app使用的接口, 接口域名：https://localhost:8888/
  title: api接口文档
  version: 0.0.1
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: params
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: '{"error_code": 0, "message": "OK", "data": {}}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login
      tags:
      - example
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"